+++ src/extension.ts
@@ -1,6 +1,7 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from "vscode";
+import { loadPostfixCmp } from "./cmp";
 import { CodeAction } from "./codeaction";
 import { setSelectionCmd, setSelectionHandle } from "./command";
 
@@ -17,6 +18,15 @@ export function activate(context: vscode.ExtensionContext) {
     vscode.commands.registerCommand(setSelectionCmd, setSelectionHandle)
   );
 
+  ["go", "javascript", "typescript", "c", "rust", "java"].forEach((lang) => {
+    context.subscriptions.push(
+      vscode.languages.registerCompletionItemProvider(
+        lang,
+        loadPostfixCmp(lang),
+        "."
+      )
+    );
+  });
 }
 
 // This method is called when your extension is deactivated